+++ src/asynchro.rs
@@ -1,25 +1,13 @@
 use std::fmt;
 use std::marker::PhantomData;
 
-use crate::asynchro_fast::{
-    interp_cubic, interp_lin, interp_quintic, interp_septic, PolynomialDegree,
-};
+use crate::asynchro_fast::{InnerPoly, PolynomialDegree};
 use crate::asynchro_sinc::{
-    interp_cubic as interp_cubic_array, interp_lin as interp_lin_array,
-    interp_quad as interp_quad_array, make_interpolator, SincInterpolationParameters,
-    SincInterpolationType,
+    make_interpolator, InnerSinc, SincInterpolationParameters, SincInterpolationType,
 };
 use crate::error::{ResampleError, ResampleResult, ResamplerConstructionError};
-use crate::interpolation::*;
 use crate::sinc_interpolator::SincInterpolator;
 use crate::{update_mask_from_buffers, validate_buffers, Fixed, Resampler, Sample};
-
-macro_rules! t {
-    // Shorter form of T::coerce(value)
-    ($expression:expr) => {
-        T::coerce($expression)
-    };
-}
 
 /// Functions for making the scalar product with a sinc.
 pub trait InnerResampler<T>: Send {
@@ -37,274 +25,6 @@ pub trait InnerResampler<T>: Send {
 
     /// Get interpolator length.
     fn len(&self) -> usize;
-}
-
-pub struct InnerSinc<T> {
-    interpolator: Box<dyn SincInterpolator<T>>,
-    interpolation: SincInterpolationType,
-}
-
-impl<T> InnerResampler<T> for InnerSinc<T>
-where
-    T: Sample,
-{
-    fn process(
-        &self,
-        idx: f64,
-        nbr_frames: usize,
-        channel_mask: &[bool],
-        t_ratio: f64,
-        t_ratio_increment: f64,
-        wave_in: &[Vec<T>],
-        wave_out: &mut [&mut [T]],
-    ) -> f64 {
-        let mut t_ratio = t_ratio;
-        let mut idx = idx;
-        let interpolator_len = self.interpolator.len();
-        match self.interpolation {
-            SincInterpolationType::Cubic => {
-                let oversampling_factor = self.interpolator.nbr_sincs();
-                let mut points = [T::zero(); 4];
-                let mut nearest = [(0isize, 0isize); 4];
-                for n in 0..nbr_frames {
-                    t_ratio += t_ratio_increment;
-                    idx += t_ratio;
-                    get_nearest_times_4(idx, oversampling_factor as isize, &mut nearest);
-                    let frac = idx * oversampling_factor as f64
-                        - (idx * oversampling_factor as f64).floor();
-                    let frac_offset = T::coerce(frac);
-                    for (chan, active) in channel_mask.iter().enumerate() {
-                        if *active {
-                            let buf = &wave_in[chan];
-                            for (n, p) in nearest.iter().zip(points.iter_mut()) {
-                                *p = self.interpolator.get_sinc_interpolated(
-                                    buf,
-                                    (n.0 + 2 * interpolator_len as isize) as usize,
-                                    n.1 as usize,
-                                );
-                            }
-                            wave_out[chan][n] = interp_cubic_array(frac_offset, &points);
-                        }
-                    }
-                }
-            }
-            SincInterpolationType::Quadratic => {
-                let oversampling_factor = self.interpolator.nbr_sincs();
-                let mut points = [T::zero(); 3];
-                let mut nearest = [(0isize, 0isize); 3];
-                for n in 0..nbr_frames {
-                    t_ratio += t_ratio_increment;
-                    idx += t_ratio;
-                    get_nearest_times_3(idx, oversampling_factor as isize, &mut nearest);
-                    let frac = idx * oversampling_factor as f64
-                        - (idx * oversampling_factor as f64).floor();
-                    let frac_offset = T::coerce(frac);
-                    for (chan, active) in channel_mask.iter().enumerate() {
-                        if *active {
-                            let buf = &wave_in[chan];
-                            for (n, p) in nearest.iter().zip(points.iter_mut()) {
-                                *p = self.interpolator.get_sinc_interpolated(
-                                    buf,
-                                    (n.0 + 2 * interpolator_len as isize) as usize,
-                                    n.1 as usize,
-                                );
-                            }
-                            wave_out[chan][n] = interp_quad_array(frac_offset, &points);
-                        }
-                    }
-                }
-            }
-            SincInterpolationType::Linear => {
-                let oversampling_factor = self.interpolator.nbr_sincs();
-                let mut points = [T::zero(); 2];
-                let mut nearest = [(0isize, 0isize); 2];
-                for n in 0..nbr_frames {
-                    t_ratio += t_ratio_increment;
-                    idx += t_ratio;
-                    get_nearest_times_2(idx, oversampling_factor as isize, &mut nearest);
-                    let frac = idx * oversampling_factor as f64
-                        - (idx * oversampling_factor as f64).floor();
-                    let frac_offset = T::coerce(frac);
-                    for (chan, active) in channel_mask.iter().enumerate() {
-                        if *active {
-                            let buf = &wave_in[chan];
-                            for (n, p) in nearest.iter().zip(points.iter_mut()) {
-                                *p = self.interpolator.get_sinc_interpolated(
-                                    buf,
-                                    (n.0 + 2 * interpolator_len as isize) as usize,
-                                    n.1 as usize,
-                                );
-                            }
-                            wave_out[chan][n] = interp_lin_array(frac_offset, &points);
-                        }
-                    }
-                }
-            }
-            SincInterpolationType::Nearest => {
-                let oversampling_factor = self.interpolator.nbr_sincs();
-                let mut point;
-                let mut nearest;
-                for n in 0..nbr_frames {
-                    t_ratio += t_ratio_increment;
-                    idx += t_ratio;
-                    nearest = get_nearest_time(idx, oversampling_factor as isize);
-                    for (chan, active) in channel_mask.iter().enumerate() {
-                        if *active {
-                            let buf = &wave_in[chan];
-                            point = self.interpolator.get_sinc_interpolated(
-                                buf,
-                                (nearest.0 + 2 * interpolator_len as isize) as usize,
-                                nearest.1 as usize,
-                            );
-                            wave_out[chan][n] = point;
-                        }
-                    }
-                }
-            }
-        }
-        idx
-    }
-
-    fn len(&self) -> usize {
-        self.interpolator.len()
-    }
-}
-
-pub struct InnerPoly<T> {
-    _phantom: PhantomData<T>,
-    interpolation: PolynomialDegree,
-}
-
-impl<T> InnerResampler<T> for InnerPoly<T>
-where
-    T: Sample,
-{
-    fn process(
-        &self,
-        idx: f64,
-        nbr_frames: usize,
-        channel_mask: &[bool],
-        t_ratio: f64,
-        t_ratio_increment: f64,
-        wave_in: &[Vec<T>],
-        wave_out: &mut [&mut [T]],
-    ) -> f64 {
-        let interpolator_len = self.len();
-        let mut t_ratio = t_ratio;
-        let mut idx = idx;
-        match self.interpolation {
-            PolynomialDegree::Septic => {
-                for n in 0..nbr_frames {
-                    t_ratio += t_ratio_increment;
-                    idx += t_ratio;
-                    let idx_floor = idx.floor();
-                    let start_idx = idx_floor as isize - 3;
-                    let frac = idx - idx_floor;
-                    let frac_offset = t!(frac);
-                    for (chan, active) in channel_mask.iter().enumerate() {
-                        if *active {
-                            unsafe {
-                                let buf = wave_in.get_unchecked(chan).get_unchecked(
-                                    (start_idx + 2 * interpolator_len as isize) as usize
-                                        ..(start_idx + 2 * interpolator_len as isize + 8) as usize,
-                                );
-                                *wave_out.get_unchecked_mut(chan).get_unchecked_mut(n) =
-                                    interp_septic(frac_offset, buf);
-                            }
-                        }
-                    }
-                }
-            }
-            PolynomialDegree::Quintic => {
-                for n in 0..nbr_frames {
-                    t_ratio += t_ratio_increment;
-                    idx += t_ratio;
-                    let idx_floor = idx.floor();
-                    let start_idx = idx_floor as isize - 2;
-                    let frac = idx - idx_floor;
-                    let frac_offset = t!(frac);
-                    for (chan, active) in channel_mask.iter().enumerate() {
-                        if *active {
-                            unsafe {
-                                let buf = wave_in.get_unchecked(chan).get_unchecked(
-                                    (start_idx + 2 * interpolator_len as isize) as usize
-                                        ..(start_idx + 2 * interpolator_len as isize + 6) as usize,
-                                );
-                                *wave_out.get_unchecked_mut(chan).get_unchecked_mut(n) =
-                                    interp_quintic(frac_offset, buf);
-                            }
-                        }
-                    }
-                }
-            }
-            PolynomialDegree::Cubic => {
-                for n in 0..nbr_frames {
-                    t_ratio += t_ratio_increment;
-                    idx += t_ratio;
-                    let idx_floor = idx.floor();
-                    let start_idx = idx_floor as isize - 1;
-                    let frac = idx - idx_floor;
-                    let frac_offset = t!(frac);
-                    for (chan, active) in channel_mask.iter().enumerate() {
-                        if *active {
-                            unsafe {
-                                let buf = wave_in.get_unchecked(chan).get_unchecked(
-                                    (start_idx + 2 * interpolator_len as isize) as usize
-                                        ..(start_idx + 2 * interpolator_len as isize + 4) as usize,
-                                );
-                                *wave_out.get_unchecked_mut(chan).get_unchecked_mut(n) =
-                                    interp_cubic(frac_offset, buf);
-                            }
-                        }
-                    }
-                }
-            }
-            PolynomialDegree::Linear => {
-                for n in 0..nbr_frames {
-                    t_ratio += t_ratio_increment;
-                    idx += t_ratio;
-                    let idx_floor = idx.floor();
-                    let start_idx = idx_floor as isize;
-                    let frac = idx - idx_floor;
-                    let frac_offset = t!(frac);
-                    for (chan, active) in channel_mask.iter().enumerate() {
-                        if *active {
-                            unsafe {
-                                let buf = wave_in.get_unchecked(chan).get_unchecked(
-                                    (start_idx + 2 * interpolator_len as isize) as usize
-                                        ..(start_idx + 2 * interpolator_len as isize + 2) as usize,
-                                );
-                                *wave_out.get_unchecked_mut(chan).get_unchecked_mut(n) =
-                                    interp_lin(frac_offset, buf);
-                            }
-                        }
-                    }
-                }
-            }
-            PolynomialDegree::Nearest => {
-                for n in 0..nbr_frames {
-                    t_ratio += t_ratio_increment;
-                    idx += t_ratio;
-                    let start_idx = idx.floor() as isize;
-                    for (chan, active) in channel_mask.iter().enumerate() {
-                        if *active {
-                            unsafe {
-                                let point = wave_in.get_unchecked(chan).get_unchecked(
-                                    (start_idx + 2 * interpolator_len as isize) as usize,
-                                );
-                                *wave_out.get_unchecked_mut(chan).get_unchecked_mut(n) = *point;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        idx
-    }
-
-    fn len(&self) -> usize {
-        self.interpolation.len()
-    }
 }
 
 /// An asynchronous resampler that either takes a fixed number of input frames,
@@ -1352,7 +1072,6 @@ mod tests {
     }
 
     // ------ Sinc tests ------
-    use super::{interp_cubic, interp_lin};
     use crate::SincInterpolationParameters;
     use crate::SincInterpolationType;
     use crate::WindowFunction;
@@ -1365,42 +1084,6 @@ mod tests {
             oversampling_factor: 16,
             window: WindowFunction::BlackmanHarris2,
         }
-    }
-
-    #[test]
-    fn int_cubic() {
-        let params = basic_params();
-        let _resampler = Async::<f64>::new_sinc(1.2, 1.0, params, 1024, 2, Fixed::Input).unwrap();
-        let yvals = [0.0f64, 2.0f64, 4.0f64, 6.0f64];
-        let interp = interp_cubic(0.5f64, &yvals);
-        assert_eq!(interp, 3.0f64);
-    }
-
-    #[test]
-    fn int_lin_32() {
-        let params = basic_params();
-        let _resampler = Async::<f32>::new_sinc(1.2, 1.0, params, 1024, 2, Fixed::Input).unwrap();
-        let yvals = [1.0f32, 5.0f32];
-        let interp = interp_lin(0.25f32, &yvals);
-        assert_eq!(interp, 2.0f32);
-    }
-
-    #[test]
-    fn int_cubic_32() {
-        let params = basic_params();
-        let _resampler = Async::<f32>::new_sinc(1.2, 1.0, params, 1024, 2, Fixed::Input).unwrap();
-        let yvals = [0.0f32, 2.0f32, 4.0f32, 6.0f32];
-        let interp = interp_cubic(0.5f32, &yvals);
-        assert_eq!(interp, 3.0f32);
-    }
-
-    #[test]
-    fn int_lin() {
-        let params = basic_params();
-        let _resampler = Async::<f64>::new_sinc(1.2, 1.0, params, 1024, 2, Fixed::Input).unwrap();
-        let yvals = [1.0f64, 5.0f64];
-        let interp = interp_lin(0.25f64, &yvals);
-        assert_eq!(interp, 2.0f64);
     }
 
     #[test]
